fix: move Shared<T> to types/mod.rs
