+++ src/instance.rs
@@ -13,11 +13,11 @@ use reqwest::Client;
 use serde::{Deserialize, Serialize};
 
 use crate::errors::ChorusResult;
-use crate::gateway::{Gateway, GatewayHandle, Shared};
+use crate::gateway::{Gateway, GatewayHandle};
 use crate::ratelimiter::ChorusRequest;
 use crate::types::types::subconfigs::limits::rates::RateLimits;
 use crate::types::{
-    GeneralConfiguration, Limit, LimitType, LimitsConfiguration, User, UserSettings,
+    GeneralConfiguration, Limit, LimitType, LimitsConfiguration, Shared, User, UserSettings,
 };
 use crate::UrlBundle;
