+++ src/types/entities/auto_moderation.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::gateway::Shared;
+use crate::types::Shared;
 #[cfg(feature = "client")]
 use crate::gateway::Updateable;
