+++ src/types/mod.rs
@@ -4,6 +4,8 @@
 
 //! All the types, entities, events and interfaces of the Spacebar API.
 
+use std::sync::{Arc, RwLock};
+
 pub use config::*;
 pub use entities::*;
 pub use errors::*;
@@ -19,3 +21,11 @@ mod events;
 mod interfaces;
 mod schema;
 mod utils;
+
+/// A type alias for [`Arc<RwLock<T>>`], used to make the public facing API concerned with
+/// Composite structs more ergonomic.
+/// ## Note
+///
+/// While `T` does not have to implement `Composite` to be used with `Shared`,
+/// the primary use of `Shared` is with types that implement `Composite`.
+pub type Shared<T> = Arc<RwLock<T>>;
